Add config parser
